backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

# Data processing
pandas==2.2.3
numpy==1.26.4
python-dateutil==2.9.0

# File handling
python-multipart==0.0.12

# API clients
httpx==0.27.0
openai==1.54.0

# Utilities
pydantic==2.9.2
pydantic-settings==2.6.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1

# Optional performance
orjson==3.10.11